+++ src/collective/cover/tests/test_contentbody_tile.py
@@ -5,6 +5,8 @@ from collective.cover.tiles.base import IPersistentCoverTile
 from collective.cover.tiles.configuration import ITilesConfigurationScreen
 from collective.cover.tiles.contentbody import ContentBodyTile
 from collective.cover.tiles.permissions import ITilesPermissions
+from plone.app.testing import setRoles
+from plone.app.testing import TEST_USER_ID
 from zope.annotation.interfaces import IAnnotations
 from zope.component import getMultiAdapter
 from zope.interface.verify import verifyClass
@@ -68,6 +70,31 @@ class ContentBodyTileTestCase(unittest.TestCase):
 
         # the body need to bring our motto
         self.assertIn('Peace of mind', rendered)
+
+    def test_render_deleted_object(self):
+        text = '<h2>Peace of mind</h2>'
+        obj = self.portal['my-news-item']
+        obj.setText(text)
+
+        self.tile.populate_with_object(obj)
+        # Delete original object
+        setRoles(self.portal, TEST_USER_ID, ['Manager'])
+        self.portal.manage_delObjects(['my-news-item', ])
+
+        rendered = self.tile()
+
+        self.assertIn('Please drag&amp;drop', rendered)
+
+    def test_render_restricted_object(self):
+        text = '<h2>Peace of mind</h2>'
+        obj = self.portal['my-news-item']
+        obj.setText(text)
+
+        self.tile.populate_with_object(obj)
+        obj.manage_permission('View', [], 0)
+        rendered = self.tile()
+
+        self.assertIn('Please drag&amp;drop', rendered)
 
     def test_delete_tile_persistent_data(self):
         permissions = getMultiAdapter(